Add EventScheduler support (#27)
* Add EventScheduler support

Schedulers can be used to schedule the events of your observation. By default, Hanson uses the `CurrentThreadScheduler`, which immediatly sends events on whatever thread it currently is. Hanson also offers the `MainThreadScheduler`, which ensures events are sent on the main thread. This is useful when observing a value that can change from a background thread and you want to do UI changes based on that value.

This introduces a breaking change to the `EventPublisher` protocol. The `eventHandlers` type has been changed to add support for event schedulers.

* Use one method with default value